Refactor: Merge all expression conversions into ExpressionConverter class
